LightCommonFile: Move file related methods to file.py
